Refactor: split the original generate_features.py and extract the parallelisation aspect into a module of it's own
